+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 
 * Config option `prefix` renamed to `messageCommandPrefix`
 * Config option `game` renamed to `activity`
+* Config option `role` renamed to `userRoleIds`. Changed from string to array of strings.
 * Docker internal volume path moved from `/usr/src/markbot/config` to `/usr/app/config`
 * Database changed from JSON files to a SQLite database. You'll need to retrain the bot to use it again.
 * The bot must be explicitly granted permission to listen to a list of channels before using it. Configure it with `/listen`.